+++ src/main/java/com/hongik/devtalk/domain/auth/dto/LogoutRequestDto.java
@@ -1,4 +1,4 @@
-package com.hongik.devtalk.controller.mainpage.dto;
+package com.hongik.devtalk.domain.auth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,8 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class MainpageImagesResponseDto {
+public class LogoutRequestDto {
     
-    private ImageInfoDto intro;
-    private ImageInfoDto previousSeminar;
+    private String userId;
 }
