+++ src/main/java/com/hongik/devtalk/domain/mainpage/dto/MainpageImagesResponseDto.java
@@ -1,4 +1,4 @@
-package com.hongik.devtalk.controller.mainpage.dto;
+package com.hongik.devtalk.domain.mainpage.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ReorderResponseDto {
+public class MainpageImagesResponseDto {
     
-    private Integer updatedCount;
+    private ImageInfoDto intro;
+    private ImageInfoDto previousSeminar;
 }
